[FIX] rb_http_handler_s->next_thread is now atomic, so many threads can call rb_http_produce with no risk.
Before of this, if two different threads call rb_http_produce, next thread could be > handler->options->connections

+++ src/rb_http_handler.h
@@ -36,7 +36,7 @@ struct rb_http_handler_s {
 	int still_running;       // NORMAL_MODE: Number of easy to be processed
 	int msgs_left;           // NORMAL_MODE
 	int left;                // NORMAL_MODE
-	int next_thread;
+	uint64_t next_thread;
 
 	struct rb_http_options_s *options; // Options
 	int thread_running;                // Keep threads running if set to 1
